+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2022-02-28
+## Version 0.04
+* Added Root CA certificate to client authentication .p12 file.
+
+## Version 0.03
+* Added Windows installation instructions for importing the certificate by using the `--windowsInstallation` argument.
+
 # 2022-02-24
 ## Version 0.02
 * Fixes - adding logic to handle if only `--companyName` argument is passed without anything else.